# scca-rest/src/main/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample.",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Sample",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "${scca.rest.context-path}/encryptKey" : {
      "get" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "List EncryptKey / 需加密的Key清单",
        "description" : "",
        "operationId" : "findAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespListEncryptKeyDto"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Add EncryptKey / 新增需加密的Key",
        "description" : "",
        "operationId" : "createEncryptKey",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EncryptKeyDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Update EncryptKey / 更新需加密的Key",
        "description" : "",
        "operationId" : "updateEncryptKey",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EncryptKeyDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Delete EncryptKey / 删除需加密的Key",
        "description" : "",
        "operationId" : "deleteEncryptKey",
        "parameters" : [ {
          "name" : "encryptKeyId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/encryptKey/batch" : {
      "post" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Add EncryptKeys / 批量新增需加密的Key",
        "description" : "",
        "operationId" : "createBatch",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/EncryptKeyDto"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Delete EncryptKeys / 批量删除需加密的Key",
        "description" : "",
        "operationId" : "deleteBatch",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/env" : {
      "get" : {
        "tags" : [ "Env MGT（环境管理）" ],
        "summary" : "List Env / 环境列表",
        "description" : "",
        "operationId" : "findAllEnv",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespListEnvDto"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Env MGT（环境管理）" ],
        "summary" : "Create Env / 创建环境",
        "description" : "",
        "operationId" : "createEnv",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EnvDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Env MGT（环境管理）" ],
        "summary" : "Update Env / 更新环境",
        "description" : "",
        "operationId" : "updateEnv",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EnvDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Env MGT（环境管理）" ],
        "summary" : "Delete Env / 删除环境",
        "description" : "",
        "operationId" : "deleteEnv",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/env/detail" : {
      "get" : {
        "tags" : [ "Env MGT（环境管理）" ],
        "summary" : "Env Detail / 环境详情",
        "description" : "",
        "operationId" : "findEnvDetail",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespEnvDto"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/envParam" : {
      "get" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "List EnvParam / 环境参数列表",
        "description" : "",
        "operationId" : "findEnvParamsByEnv",
        "parameters" : [ {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespListEnvParamDto"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Add EnvParam / 添加一个环境参数",
        "description" : "",
        "operationId" : "addEnvParam",
        "parameters" : [ {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EnvParamDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Update EnvParam / 更新某个环境参数",
        "description" : "",
        "operationId" : "updateEnvParam",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EnvParamDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Delete EnvParam / 删除某个环境参数",
        "description" : "",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/envParam/batch" : {
      "post" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Add EnvParams / 新增一批环境参数",
        "description" : "",
        "operationId" : "createBatch",
        "parameters" : [ {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/EnvParamDto"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Delete EnvParams / 删除一批环境参数",
        "description" : "",
        "operationId" : "deleteBatch",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "EncryptKeyDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "ekey" : {
          "type" : "string"
        },
        "eKey" : {
          "type" : "string"
        }
      }
    },
    "EnvDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "registryAddress" : {
          "type" : "string"
        },
        "configServerName" : {
          "type" : "string"
        },
        "contextPath" : {
          "type" : "string"
        }
      }
    },
    "EnvParamDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "pkey" : {
          "type" : "string"
        },
        "pvalue" : {
          "type" : "string"
        },
        "pKey" : {
          "type" : "string"
        },
        "pValue" : {
          "type" : "string"
        }
      }
    },
...
    "WebRespEnvDto" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "$ref" : "#/definitions/EnvDto"
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
    "WebRespListEncryptKeyDto" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/EncryptKeyDto"
          }
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
    "WebRespListEnvDto" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/EnvDto"
          }
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
    "WebRespListEnvParamDto" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/EnvParamDto"
          }
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
    "WebRespString" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
  }
}